Conflicts:
	engine/engine/data/meshes/Manillar_S.mesh
	engine/engine/data/meshes/Object001.mesh
	engine/engine/data/meshes/Object002.mesh
	engine/engine/data/meshes/cartel_c.mesh
	engine/engine/data/meshes/cartel_theatre_base.mesh
	engine/engine/data/meshes/cartel_theatre_rope.mesh
	engine/engine/data/meshes/cartel_theatre_top.mesh
	engine/engine/data/meshes/colliders/esc1_puert_suelo_pared_0m_1m_b_collider.mesh
	engine/engine/data/meshes/colliders/esc1_rampa_rampa_45g_doble_a_collider.mesh
	engine/engine/data/meshes/colliders/tra1_arqui_ambos_pared_half_i_collider.mesh
	engine/engine/data/meshes/colliders/tra1_arqui_pared_esqui_in_half_a_collider.mesh
	engine/engine/data/meshes/colliders/tra1_arqui_suelo_pared_half_a_collider.mesh
	engine/engine/data/meshes/colliders/tra2_puert_suelo_pared_0m_2m_b_collider.mesh
	engine/engine/data/meshes/esc1_puert_suelo_pared_0m_1m_b.mesh
	engine/engine/data/meshes/esc1_rampa_rampa_45g_doble_a.mesh
	engine/engine/data/meshes/flecha_dech.mesh
	engine/engine/data/meshes/foco_basico_d.mesh
	engine/engine/data/meshes/lampara_steam_verde_S.mesh
	engine/engine/data/meshes/silla_camerino_base_S.mesh
	engine/engine/data/meshes/tablon_e.mesh
	engine/engine/data/meshes/tablon_largo_b.mesh
	engine/engine/data/meshes/tela_techo_tira_b.mesh
	engine/engine/data/meshes/tra1_arqui_ambos_pared_half_i.mesh
	engine/engine/data/meshes/tra1_arqui_suelo_pared_half_a.mesh
	engine/engine/data/meshes/tra2_puert_suelo_pared_0m_2m_b.mesh
	engine/engine/data/meshes/violin baja.mesh

+++ engine/engine/components/comp_audio_source.cpp
@@ -40,7 +40,7 @@ void TCompAudioSource::loadFromAtts(const std::string& elem, MKeyValue &atts) {
 
 	if (elem=="fin_param"){
 		if (!params_value.empty()){
-			const int size = params_value.size();
+			const int size = (int)params_value.size();
 			std::vector<CSoundManager::SoundParameter> params;
 
 			for (int j = 0; j < size; ++j){
